# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface TripPreviewInput {
  /** @minLength 2 */
  country: string;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  surname?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  birthYear?: number | null;
  /** @nullable */
  deathYear?: number | null;
  /** @minLength 2 */
  destination: string;
  /**
     * @minimum 1
     * @maximum 30
     */
  tripLength: number;
  budget: string;
  /** @minItems 1 */
  interests: string[];
}

export interface TripPlace {
  name: string;
  type: string;
  description: string;
  certainty: string;
  /** @nullable */
  sourceUrl: string | null;
}

export interface ItineraryItem {
  time: string;
  title: string;
  description: string;
  category: string;
}

export type EvidenceBoundaryUserProvidedFacts = { [key: string]: unknown };

export interface EvidenceBoundary {
  userProvidedFacts: EvidenceBoundaryUserProvidedFacts;
  aiGenerated: boolean;
  uncertainties: string[];
}

export interface TripPreview {
  id: string;
  title: string;
  locationLabel: string;
  interpretation: string;
  places: TripPlace[];
  sampleDay: ItineraryItem[];
  historicalContext: string;
  researchChecklist: string[];
  disclaimer: string;
  tripLength: number;
  destination: string;
  generatedAt: string;
  evidenceBoundary: EvidenceBoundary;
}

export type CheckoutInputProduct = typeof CheckoutInputProduct[keyof typeof CheckoutInputProduct];


export const CheckoutInputProduct = {
  'heritage-trip': 'heritage-trip',
  'deep-heritage-trip': 'deep-heritage-trip',
} as const;

export interface CheckoutInput {
  previewId: string;
  product: CheckoutInputProduct;
  email: string;
}

export interface CheckoutSession {
  checkoutUrl: string;
  sessionId: string;
  product: string;
  price: number;
  demoMode: boolean;
}

export interface CheckoutVerification {
  verified: boolean;
  status: string;
  /** @nullable */
  reportToken: string | null;
}

export interface ReportSection {
  title: string;
  body: string;
  items: string[];
}

export interface PaidReport {
  id: string;
  token: string;
  product: string;
  title: string;
  status: string;
  preview: TripPreview;
  sections: ReportSection[];
  evidenceBoundary: EvidenceBoundary;
  createdAt: string;
}

export interface Article {
  slug: string;
  title: string;
  excerpt: string;
  category: string;
  readTime: string;
  publishedAt: string;
}

export interface SupportTicketInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 10 */
  message: string;
}

export interface SupportTicket {
  id: string;
  status: string;
  createdAt: string;
}

export interface MetricItem {
  label: string;
  value: number;
}

export interface BriefingItem {
  label: string;
  body: string;
  tone: string;
}

export interface AdminSummary {
  revenue: number;
  orders: number;
  conversionRate: number;
  visitors: number;
  previews: number;
  failedReports: number;
  topSources: MetricItem[];
  topDestinations: MetricItem[];
  briefing: BriefingItem[];
}

export type VerifyCheckoutParams = {
sessionId: string;
};