# lib/api-zod/src/generated/types/checkoutVerification.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutVerification {
  verified: boolean;
  status: string;
  /** @nullable */
  reportToken: string | null;
}