# lib/api-zod/src/generated/types/evidenceBoundary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EvidenceBoundaryUserProvidedFacts } from './evidenceBoundaryUserProvidedFacts';

export interface EvidenceBoundary {
  userProvidedFacts: EvidenceBoundaryUserProvidedFacts;
  aiGenerated: boolean;
  uncertainties: string[];
}