# lib/api-zod/src/generated/types/paidReport.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EvidenceBoundary } from './evidenceBoundary';
import type { ReportSection } from './reportSection';
import type { TripPreview } from './tripPreview';

export interface PaidReport {
  id: string;
  token: string;
  product: string;
  title: string;
  status: string;
  preview: TripPreview;
  sections: ReportSection[];
  evidenceBoundary: EvidenceBoundary;
  createdAt: string;
}