# lib/api-zod/src/generated/types/tripPreview.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EvidenceBoundary } from './evidenceBoundary';
import type { ItineraryItem } from './itineraryItem';
import type { TripPlace } from './tripPlace';

export interface TripPreview {
  id: string;
  title: string;
  locationLabel: string;
  interpretation: string;
  places: TripPlace[];
  sampleDay: ItineraryItem[];
  historicalContext: string;
  researchChecklist: string[];
  disclaimer: string;
  tripLength: number;
  destination: string;
  generatedAt: string;
  evidenceBoundary: EvidenceBoundary;
}